+++ aws_emr_launch/constructs/step_functions/emr_launch_function.py
@@ -93,6 +93,9 @@ class EMRLaunchFunction(core.Construct):
             error='Failed to Launch Cluster',
             cause='See Execution Event "FailStateEntered" for complete error cause')
 
+        secrets = cluster_configuration.secure_configurations.values() \
+            if cluster_configuration.secure_configurations else None
+
         # Create Task for loading the cluster configuration from Parameter Store
         load_cluster_configuration = emr_tasks.LoadClusterConfigurationBuilder.build(
             self, 'LoadClusterConfigurationTask',
@@ -101,7 +104,8 @@ class EMRLaunchFunction(core.Construct):
             profile_namespace=emr_profile.namespace,
             profile_name=emr_profile.profile_name,
             configuration_namespace=cluster_configuration.namespace,
-            configuration_name=cluster_configuration.configuration_name)
+            configuration_name=cluster_configuration.configuration_name,
+            secrets=secrets)
         load_cluster_configuration.add_catch(fail, errors=['States.ALL'], result_path='$.Error')
 
         # Create Task for overriding cluster configurations