+++ aws_emr_launch/lambda_sources/emr_utilities/load_cluster_configuration.py
@@ -12,8 +12,9 @@
 # permissions and limitations under the License.
 
 import os
-import boto3
 import json
+import boto3
+import base64
 
 from logzero import logger
 from typing import Dict, List
@@ -25,6 +26,7 @@ PROFILES_SSM_PARAMETER_PREFIX = '/emr_launch/emr_profiles'
 CONFIGURATIONS_SSM_PARAMETER_PREFIX = '/emr_launch/cluster_configurations'
 
 ssm = boto3.client('ssm')
+secretsmanager = boto3.client('secretsmanager')
 
 
 class EMRProfileNotFoundError(Exception):
@@ -35,10 +37,38 @@ class ClusterConfigurationNotFoundError(Exception):
     pass
 
 
+class SecretNotFoundError(Exception):
+    pass
+
+
+class SecretDecryptionFailureError(Exception):
+    pass
+
+
 def get_parameter_value(ssm_parameter_prefix: str, name: str, namespace: str = 'default'):
     configuration_json = ssm.get_parameter(
         Name=f'{ssm_parameter_prefix}/{namespace}/{name}')['Parameter']['Value']
     return json.loads(configuration_json)
+
+
+def get_secret_value(secret_id: str):
+    try:
+        secret_response = secretsmanager.get_secret_value(
+            SecretId=secret_id
+        )
+    except ClientError as e:
+        if e.response['Error']['Code'] == 'DecryptionFailureException':
+            raise SecretDecryptionFailureError(f'SecretDecryptionFailure: {secret_id}')
+        elif e.response['Error']['Code'] == 'ResourceNotFoundException':
+            raise SecretNotFoundError(f'SecretNotFound: {secret_id}')
+        else:
+            raise e
+
+    val = json.loads(secret_response.pop('SecretString')) \
+        if 'SecretString' in secret_response \
+        else json.loads(base64.b64decode(secret_response.pop('SecretBinary')))
+    logger.info(f'SecretFound: {json.dumps(val)}')
+    return val
 
 
 def log_and_raise(e, event):
@@ -94,7 +124,7 @@ def handler(event, context):
         cluster_configuration = get_parameter_value(
             ssm_parameter_prefix=CONFIGURATIONS_SSM_PARAMETER_PREFIX,
             namespace=configuration_namespace,
-            name=configuration_name)['ClusterConfiguration']
+            name=configuration_name)
         logger.info(f'ConfigurationFound: {json.dumps(cluster_configuration)}')
     except ClientError as e:
         if e.response['Error']['Code'] == 'ParameterNotFound':
@@ -106,6 +136,15 @@ def handler(event, context):
     try:
         logs_bucket = emr_profile.get('LogsBucket', None)
         logs_path = emr_profile.get('LogsPath', '')
+
+        secure_configurations = cluster_configuration.get('SecureConfigurations', None)
+        cluster_configuration = cluster_configuration['ClusterConfiguration']
+
+        if secure_configurations:
+            for classification, secret_id in secure_configurations.items():
+                properties = get_secret_value(secret_id)
+                cluster_configuration['Configurations'] = update_configurations(
+                    cluster_configuration['Configurations'], classification, properties)
 
         cluster_configuration['Name'] = cluster_name
         cluster_configuration['LogUri'] = \