+++ examples/cluster_configurations/app.py
@@ -5,6 +5,7 @@ import os
 from aws_cdk import (
     aws_ec2 as ec2,
     aws_s3 as s3,
+    aws_secretsmanager as secretsmanager,
     core
 )
 
@@ -39,6 +40,10 @@ bootstrap = emr_code.EMRBootstrapAction(
 # Cluster Configurations that use InstanceGroups are deployed to a Private subnet
 subnet = vpc.private_subnets[0]
 
+# Load a SecretsManger Secret with secure RDS Metastore credentials
+secret = secretsmanager.Secret.from_secret_arn(
+    stack, 'Secret', os.environ['EMR_LAUNCH_EXAMPLES_SECURE_CONFIGS'])
+
 # Create a basic Cluster Configuration using InstanceGroups, the Subnet and Bootstrap
 # Action defined above, the EMR Profile we loaded, and defaults defined in
 # the InstanceGroupConfiguration
@@ -67,6 +72,7 @@ high_mem_cluster_config = cluster_configuration.InstanceGroupConfiguration(
     bootstrap_actions=[bootstrap],
     step_concurrency_level=5,
     core_instance_type='r5.2xlarge',
-    core_instance_count=2)
+    core_instance_count=2,
+    secure_configurations={'hive-site': secret})
 
 app.synth()