+++ examples/cluster_configurations/bootstrap_source/test_bootstrap.sh
@@ -1,2 +1,3 @@
 #!/bin/bash
 echo "hello world" && echo "Args: $@" && exit 0
+pip install boto3