+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changelog for emr-launch
 1.0.0 (unreleased)
 ------------------
 
+- Secure cluster Configurations loaded from SecretsManager
+
 - Helper functions to add spark packages and jars
 
 - implement EMR/Step Functions integrations